Improved simplicity and dataTypes
