+++ src/lib.rs
@@ -1,4 +1,3 @@
-
 //! # Atelier
 //! [`Atelier`] provides a computational framework that can be used as a workshop to develop Financial Machine Learning models, especifically battle tested to the level of Market microstructure.
 //! Features include:
@@ -7,23 +6,22 @@
 //!
 //! ## Example: Synthetic Orderbook
 //! ```
-//! use atelier::data::market::orderbook::Orderbook;
-//! 
+//! use atelier::data::market::Orderbook;
+//!
 //! let bid_price = 50_000.00;
 //! let ask_price = 50_100.00;
 //! let tick_size = 100.0;
 //! let n_levels = 200;
 //! let n_orders = 300;
-//! 
-//! let i_ob = Orderbook::synthetize(bid_price, 
-//!        ask_price, tick_size, 
-//!        n_levels, n_orders);
+//!
+//! let i_ob = Orderbook::synthetize(bid_price, ask_price,
+//!        tick_size, n_levels, n_orders);
 //! ```
 
 /// Definition of data structures and core datatypes
 pub mod data;
 
-/// Tools to create generators of data from specification 
+/// Tools to create generators of data from specification
 /// of parametric probability distributions.
 pub mod simulation;
