+++ src/simulation/randomizer.rs
@@ -1,4 +1,4 @@
-use crate::data::market::order::Order;
+use crate::data::market::{Order, OrderType, Side};
 use rand::distributions::Uniform;
 use rand::Rng;
 use std::time::{SystemTime, UNIX_EPOCH};
@@ -6,7 +6,7 @@ use std::time::{SystemTime, UNIX_EPOCH};
 // Create a random Order according to the provided parameters
 // it uses structs::marketdata::order::Order
 
-pub fn randomize_order(side: String, price: f64) -> Order {
+pub fn randomize_order(side: Side, price: f64, order_type: OrderType) -> Order {
     let mut uni_rand = rand::thread_rng();
 
     // Randomize order_ts
@@ -32,9 +32,13 @@ pub fn randomize_order(side: String, price: f64) -> Order {
     // Randomize order_id
     let order_id: u32 = 123;
 
+    // Parse Order Type
+    let order_type = order_type;
+
     Order {
         order_id,
         order_ts,
+        order_type,
         side,
         price,
         amount,
